feature: opt-in to write index and scene number to scene files (#288)
* write basic scene index to note on draft change

* sync scene indices when setting is enabled

* write scene number to file

* persist "writeProperty" setting

* inform users that two properties will be written to scenes

* check opt-in during scene creation

not sure how this got removed during the rebase, but this commit adds the check back in so that the scene index is not written to the new file if the user has not opted-in.

* move writting scene number to draft utils

it made more sense for the writing of draft information about the scene to be within the "createScene" usecase code than the note utils file.

createNoteWithPotentialTemplate now returns the TFile, if successful.  Updated documentation to reflect this new behavior.